Moves sv-variant to server/variant
