+++ scout/server/blueprints/variant/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .views import variant_bp
