+++ scout/server/blueprints/variant/templates/variant/sv-variant.html
@@ -614,7 +614,7 @@
 <ul class="list-group">
   <div class="list-group-item">
     <strong>Case {{ case.display_name }}</strong><br>
-    <a href="{{ url_for('variants.sv_variant', institute_id=institute._id,
+    <a href="{{ url_for('variant.sv_variant', institute_id=institute._id,
                         case_name=case.display_name, variant_id=variant._id) }}">
       {% if variant.end_chrom and variant.end_chrom == variant.chromosome %}
         {{variant.chromosome}}:{{variant.position}}-{{variant.end}}
